chore: generate essential directories at run-time
